requirements.txt
numpy==2.1.1
psycopg2==2.9.9
python-dotenv==1.0.1
tqdm==4.66.5